# Docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sky Template API
  version: 1.0.0
  description: Generated from docs/source/api.md and controllers
  x-generated-from:
    - docs/source/api.md
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Cookies
  schemas:
    AuthRequest:
      type: object
      description: TODO
    AuthResponse:
      type: object
      description: TODO
...
paths:
  /api/auth/sign-in:
    post:
      tags: [Shared]
      summary: Sign in
      description: User login
      operationId: SharedAuthController_post_sign_in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...